+++ guppylang-internals/src/guppylang_internals/decorator/__init__.py
@@ -1,0 +1,12 @@
+from .custom import custom_function, custom_type, hugr_op
+from .ty import extend_type
+from .wasm import wasm, wasm_module
+
+__all__ = [
+    "custom_function",
+    "custom_type",
+    "extend_type",
+    "hugr_op",
+    "wasm",
+    "wasm_module",
+]
