--- guppylang-internals/src/guppylang_internals/std/wasm/__init__.py
@@ -1,6 +1,0 @@
-from .decorator import wasm, wasm_module
-
-__all__ = [
-    "wasm",
-    "wasm_module",
-]
