+++ tests/error/wasm_errors/wasm_bad_input.py
@@ -1,5 +1,5 @@
 from guppylang import guppy
-from guppylang_internals.std.wasm import wasm, wasm_module
+from guppylang_internals.decorator import wasm, wasm_module
 
 from tests.util import get_wasm_file
