+++ tests/error/wasm_errors/wasm_fn_without_module1.py
@@ -1,5 +1,5 @@
 from guppylang import guppy
-from guppylang_internals.std.wasm import wasm
+from guppylang_internals.decorator.wasm import wasm
 
 @guppy.struct
 class Foo:
