+++ tests/error/wasm_errors/wasm_without_args.py
@@ -1,5 +1,5 @@
 from guppylang import guppy
-from guppylang_internals.std.wasm import wasm, wasm_module
+from guppylang_internals.decorator.wasm import wasm, wasm_module
 
 from tests.util import get_wasm_file
